Beginning of construction of pie chart
Started construction on pie chart. Basic ideas created and functions are effectively coordinated. Needs work on overlapping values of the chart and the immense size of the "other" category if a variety of tasks are being completed simultaneously. The pie chart takes matplotlib.plt priority which causes the column graph to bug out and become increasingly skinny and compressed. Consider creating a secondary matplotlib file to run in tandem with the main Appreader program and generate the pie chart in a new environment that avoids interaction with the main column graph.