Main Dashboard + General Edits
Dashboard adjusted to be a central hub instead of a login location - login idea deprecated due to incompatibility with design. Basic idea for dashboard is to be a quick opening location where the user can quickly check the next task on their checklist and see their daily activity as opposed to full activity for all time use. Consider using datetime library to check the country's date origins and saving daily time data and graph to a seperate json file - maybe localise the daily time to an entirely seperate python/json pair?