MASSIVE BUGFIX
Patched an omnipresent and extremely difficult bug wherein the app-time calculations would add additional time to the running programs if the python script for appreader_remake_testing was shut down, calculating from the original starting time. Next step is to fix and improve the daily time data dashboard as it currently only displays the basic checklist. Could further expand this into a weekly time data segment to be used in page 1.