Dailytimedata dictionary + json file
Now records daily time data for apps that resets each day, using the datetime.today function to determine the current integer of the day, and resetting the json file whenever the day value changes. Next step is to integrate the data into a graph and display on the dashboard page.